correcting inheritance and instantiations of non-perishables
in the food class i should've been inheriting from the items class and nothing else so in the next version i intend to potentially move maturationStr elsewhere. i also added a method which counts the instantiations of child classes which only works under the pretense that getgenerationcount is called. i found a way to make weeds have an infinite value for its death age and maturation age.